chore: flatten src/api
