--- frontend/src/api/admin.ts
@@ -1,121 +1,0 @@
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { AxiosResponse } from "axios";
-import { httpClient } from "./httpClient";
-import { AppLanguage, DashboardResponse, Gender, Race } from "./user";
-
-export interface CreateUserRequest {
-  contactNumber: number;
-  name: string;
-  age: number;
-  alias: string;
-  race: Race;
-  appLanguage: AppLanguage;
-  gender: Gender;
-  postalCode: number;
-  floor: number;
-  block: string;
-  unit: string;
-}
-
-export interface SignUpAdminRequest {
-  clerk_id: string;
-  contactNumber: number;
-}
-
-export interface UpdateUserRequest extends CreateUserRequest {}
-
-export async function getSignUpAdminResponse(
-  signUpAdminRequest: SignUpAdminRequest
-): Promise<AxiosResponse<null>> {
-  return await httpClient.post("/admin/sign-up", signUpAdminRequest);
-}
-
-export function useGetSignUpAdminResponse() {
-  return useMutation({
-    mutationFn: (request: SignUpAdminRequest) =>
-      getSignUpAdminResponse(request),
-  });
-}
-
-export async function getCreateUserResponse(
-  createUserRequest: CreateUserRequest
-): Promise<AxiosResponse<null>> {
-  return await httpClient.post("/user", createUserRequest);
-}
-
-export function useGetCreateNewUser() {
-  return useMutation({
-    mutationFn: (request: CreateUserRequest) => getCreateUserResponse(request),
-  });
-}
-export async function getAdminUserResponse(
-  userId: number
-): Promise<DashboardResponse> {
-  const response = await httpClient.get(`/user/${userId}`);
-  return response.data;
-}
-
-export function useGetAdminUserResponse(userId: number) {
-  return useQuery({
-    queryKey: ["getAdminUserResponse", userId],
-    queryFn: () => getAdminUserResponse(userId),
-    enabled: !!userId,
-  });
-}
-
-export async function getDeleteUserResponse(userId: number): Promise<null> {
-  const response = await httpClient.delete(`/user/${userId}`);
-  return response.data;
-}
-
-export function useGetDeleteUserResponse() {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: (userId: number) => getDeleteUserResponse(userId),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["getAdminDashboardResponse"] });
-    },
-  });
-}
-
-export async function getUpdateUserResponse(
-  userId: number,
-  createUserRequest: CreateUserRequest
-): Promise<null> {
-  const response = await httpClient.put(`/user/${userId}`, createUserRequest);
-  return response.data;
-}
-
-export function useUpdateUser() {
-  return useMutation({
-    mutationFn: ({
-      userId,
-      request,
-    }: {
-      userId: number;
-      request: CreateUserRequest;
-    }) => getUpdateUserResponse(userId, request),
-  });
-}
-
-export async function getSuspendUserResponse(userId: number): Promise<null> {
-  const response = await httpClient.put(`/user/${userId}/suspend`);
-  return response.data;
-}
-
-export function useSuspendUser() {
-  return useMutation({
-    mutationFn: (userId: number) => getSuspendUserResponse(userId),
-  });
-}
-
-export async function getUnsuspendUserResponse(userId: number): Promise<null> {
-  const response = await httpClient.put(`/user/${userId}/unsuspend`);
-  return response.data;
-}
-
-export function useUnsuspendUser() {
-  return useMutation({
-    mutationFn: (userId: number) => getUnsuspendUserResponse(userId),
-  });
-}
