+++ frontend/src/pages/HomePage/MoodBtns/index.tsx
@@ -1,7 +1,7 @@
 import { Box, SlideFade, useDisclosure } from "@chakra-ui/react";
 import moment, { Moment } from "moment";
 import { useState } from "react";
-import { MoodValue } from "../../../api/user";
+import { MoodValue } from "../../../api/types";
 import ModalMoodStreak from "../../../components/ModalMoodStreak";
 import MoodBtn from "../../../components/MoodBtn";
 import MoodMessage from "../../../components/MoodMessage";
